optimize file receive, add pin, add key deriving using pin

+++ lib/ui/pin/pin_screen.dart
@@ -1,6 +1,7 @@
 import 'package:encrypt_shared_preferences/provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:safe_file_sender/crypto/crypto_core.dart';
 import 'package:safe_file_sender/dev/logger.dart';
 import 'package:safe_file_sender/utils/navigator_utils.dart';
 import 'package:safe_file_sender/cache/preferences_cache_keys.dart';
@@ -8,10 +9,16 @@ import 'package:safe_file_sender/models/path_values.dart';
 import 'package:safe_file_sender/ui/widgets/common_inherited_widget.dart';
 import 'package:safe_file_sender/utils/context_utils.dart';
 
-class PinScreen extends StatelessWidget {
-  final TextEditingController _textEditingController = TextEditingController();
-
+class PinScreen extends StatefulWidget {
   PinScreen({super.key});
+
+  @override
+  State<PinScreen> createState() => _PinScreenState();
+}
+
+class _PinScreenState extends State<PinScreen> {
+  final TextEditingController _textEditingController = TextEditingController();
+  bool _loading = false;
 
   @override
   Widget build(BuildContext context) {
@@ -39,11 +46,17 @@ class PinScreen extends StatelessWidget {
                     onPressed: () {
                       _handleTap(context);
                     },
-                    child: context.preferences
-                                .getString(PreferencesCacheKeys.pin) !=
-                            null
-                        ? Text(context.localization.done)
-                        : Text(context.localization.setupPin),
+                    child: _loading
+                        ? const SizedBox(
+                            width: 16,
+                            height: 16,
+                            child: CircularProgressIndicator(),
+                          )
+                        : context.preferences
+                                    .getString(PreferencesCacheKeys.pin) !=
+                                null
+                            ? Text(context.localization.done)
+                            : Text(context.localization.setupPin),
                   ),
                 ),
               ],
@@ -77,8 +90,20 @@ class PinScreen extends StatelessWidget {
     }
   }
 
-  _goHome(BuildContext context) {
-    Navigator.pushNamedAndRemoveUntil(
-        context, PathValues.home, NavigatorRoutePredicates.deleteAll);
+  _goHome(BuildContext context) async {
+    setState(() {
+      _loading = true;
+    });
+    final key =
+        await AppCrypto.deriveKeyIsolate(_textEditingController.text.trim());
+    setState(() {
+      _loading = false;
+    });
+    if (context.mounted) {
+      logMessage('Key derived : $key');
+      context.appTempData.setPinDerivedKey(key);
+      Navigator.pushNamedAndRemoveUntil(
+          context, PathValues.home, NavigatorRoutePredicates.deleteAll);
+    }
   }
 }
